chore: remove wcv1 to dApps (#5316)
