+++ src/AppProviders.tsx
@@ -5,8 +5,7 @@ import {
   createStandaloneToast,
 } from '@chakra-ui/react'
 import { DefiManagerProvider } from 'features/defi/contexts/DefiManagerProvider/DefiManagerProvider'
-import { WalletConnectBridgeProvider } from 'plugins/walletConnectToDapps/v1/WalletConnectBridgeProvider'
-import { WalletConnectV2Provider } from 'plugins/walletConnectToDapps/v2/WalletConnectV2Provider'
+import { WalletConnectV2Provider } from 'plugins/walletConnectToDapps/WalletConnectV2Provider'
 import React, { useCallback } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
 import { HelmetProvider } from 'react-helmet-async'
@@ -39,9 +38,17 @@ const manager = createLocalStorageManager('ss-theme')
 
 export function AppProviders({ children }: ProvidersProps) {
   const { ToastContainer } = createStandaloneToast()
-  const handleError = useCallback((error: Error, info: { componentStack: string }) => {
-    getMixPanel()?.track(MixPanelEvents.Error, { error, info })
-  }, [])
+  const handleError = useCallback(
+    (
+      error: Error,
+      info: {
+        componentStack: string
+      },
+    ) => {
+      getMixPanel()?.track(MixPanelEvents.Error, { error, info })
+    },
+    [],
+  )
   return (
     <HelmetProvider>
       <ReduxProvider store={store}>
@@ -56,23 +63,21 @@ export function AppProviders({ children }: ProvidersProps) {
                 <BrowserRouterProvider>
                   <I18nProvider>
                     <WalletProvider>
-                      <WalletConnectBridgeProvider>
-                        <WalletConnectV2Provider>
-                          <KeepKeyProvider>
-                            <ErrorBoundary FallbackComponent={ErrorPage} onError={handleError}>
-                              <ModalProvider>
-                                <TransactionsProvider>
-                                  <AppProvider>
-                                    <FoxEthProvider>
-                                      <DefiManagerProvider>{children}</DefiManagerProvider>
-                                    </FoxEthProvider>
-                                  </AppProvider>
-                                </TransactionsProvider>
-                              </ModalProvider>
-                            </ErrorBoundary>
-                          </KeepKeyProvider>
-                        </WalletConnectV2Provider>
-                      </WalletConnectBridgeProvider>
+                      <WalletConnectV2Provider>
+                        <KeepKeyProvider>
+                          <ErrorBoundary FallbackComponent={ErrorPage} onError={handleError}>
+                            <ModalProvider>
+                              <TransactionsProvider>
+                                <AppProvider>
+                                  <FoxEthProvider>
+                                    <DefiManagerProvider>{children}</DefiManagerProvider>
+                                  </FoxEthProvider>
+                                </AppProvider>
+                              </TransactionsProvider>
+                            </ModalProvider>
+                          </ErrorBoundary>
+                        </KeepKeyProvider>
+                      </WalletConnectV2Provider>
                     </WalletProvider>
                   </I18nProvider>
                 </BrowserRouterProvider>
