+++ src/plugins/walletConnectToDapps/components/header/DappHeaderMenuSummaryV2.tsx
@@ -1,6 +1,6 @@
 import { Accordion, MenuDivider, MenuGroup } from '@chakra-ui/react'
 import type { SessionTypes } from '@walletconnect/types'
-import { useWalletConnectV2 } from 'plugins/walletConnectToDapps/v2/WalletConnectV2Provider'
+import { useWalletConnectV2 } from 'plugins/walletConnectToDapps/WalletConnectV2Provider'
 import { useMemo } from 'react'
 import { useTranslate } from 'react-polyglot'
