+++ src/plugins/walletConnectToDapps/components/header/Session.tsx
@@ -12,9 +12,9 @@ import {
 import type { SessionTypes } from '@walletconnect/types'
 import { getSdkError } from '@walletconnect/utils'
 import dayjs from 'dayjs'
+import { WalletConnectActionType } from 'plugins/walletConnectToDapps/types'
 import { extractConnectedAccounts } from 'plugins/walletConnectToDapps/utils'
-import { WalletConnectActionType } from 'plugins/walletConnectToDapps/v2/types'
-import { useWalletConnectV2 } from 'plugins/walletConnectToDapps/v2/WalletConnectV2Provider'
+import { useWalletConnectV2 } from 'plugins/walletConnectToDapps/WalletConnectV2Provider'
 import { useCallback, useMemo } from 'react'
 import { RawText, Text } from 'components/Text'
 import { selectSelectedLocale } from 'state/slices/selectors'
