+++ src/plugins/walletConnectToDapps/components/modals/connect/ConnectContent.tsx
@@ -11,24 +11,14 @@ import {
   Link,
   VStack,
 } from '@chakra-ui/react'
-import type { AssetId } from '@shapeshiftoss/caip'
-import { ethAssetId } from '@shapeshiftoss/caip'
-import {
-  isValidWalletConnectUri,
-  isWalletConnectV1Uri,
-  isWalletConnectV2Uri,
-} from 'plugins/walletConnectToDapps/components/modals/connect/utils'
-import { useWalletConnect } from 'plugins/walletConnectToDapps/v1/WalletConnectBridgeContext'
-import { useCallback, useMemo, useState } from 'react'
+import { isWalletConnectV2Uri } from 'plugins/walletConnectToDapps/components/modals/connect/utils'
+import { useCallback, useState } from 'react'
 import { useForm, useWatch } from 'react-hook-form'
 import { useTranslate } from 'react-polyglot'
-import { AccountDropdown } from 'components/AccountDropdown/AccountDropdown'
 import { QRCodeIcon } from 'components/Icons/QRCode'
 import { WalletConnectIcon } from 'components/Icons/WalletConnectIcon'
 import { QrCodeScanner } from 'components/QrCodeScanner/QrCodeScanner'
 import { Text } from 'components/Text'
-import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
-import { useFeatureFlag } from 'hooks/useFeatureFlag/useFeatureFlag'
 
 type FormValues = {
   uri: string
@@ -45,10 +35,6 @@ export const ConnectContent: React.FC<ConnectContentProps> = ({
   const translate = useTranslate()
   const [isQrCodeView, setIsQrCodeView] = useState<boolean>(false)
   const toggleQrCodeView = useCallback(() => setIsQrCodeView(v => !v), [])
-  const { evmChainId, setWcAccountId } = useWalletConnect()
-
-  const isWalletConnectToDappsV1Enabled = useFeatureFlag('WalletConnectToDapps')
-  const isWalletConnectToDappsV2Enabled = useFeatureFlag('WalletConnectToDappsV2')
 
   const handleForm = (values: FormValues) => handleConnect(values.uri)
 
@@ -66,16 +52,7 @@ export const ConnectContent: React.FC<ConnectContentProps> = ({
   )
 
   const uri = useWatch({ control, name: 'uri' })
-  const isWalletConnectV1 = isWalletConnectV1Uri(uri)
-  const isWalletConnectV2 = isWalletConnectV2Uri(uri)
-  const isValidUri = isValidWalletConnectUri(uri)
-
-  const feeAssetId: AssetId = useMemo(() => {
-    if (!evmChainId) return ethAssetId
-    const chainAdapter = getChainAdapterManager().get(evmChainId)
-    if (!chainAdapter) return ethAssetId
-    return chainAdapter.getFeeAssetId()
-  }, [evmChainId])
+  const isValidUri = isWalletConnectV2Uri(uri)
 
   if (isQrCodeView)
     return <QrCodeScanner onSuccess={handleQrScanSuccess} onBack={toggleQrCodeView} />
@@ -85,10 +62,6 @@ export const ConnectContent: React.FC<ConnectContentProps> = ({
     switch (true) {
       case uri === '':
         return `${commonString}.connect`
-      case isWalletConnectV1 && !isWalletConnectToDappsV1Enabled:
-        return `${commonString}.v1UriUnsupported`
-      case isWalletConnectV2 && !isWalletConnectToDappsV2Enabled:
-        return `${commonString}.v2UriUnsupported`
       case !isValidUri:
         return `${commonString}.invalidUri`
       default:
@@ -111,13 +84,6 @@ export const ConnectContent: React.FC<ConnectContentProps> = ({
           </Link>
 
           <FormControl isInvalid={Boolean(formState.errors.uri)} mb={6}>
-            {!isWalletConnectV2 && (
-              <AccountDropdown
-                buttonProps={{ width: 'full', ml: 0, mr: 0 }}
-                assetId={feeAssetId}
-                onChange={setWcAccountId}
-              />
-            )}
             <InputGroup size='lg'>
               <InputRightElement>
                 <IconButton
