+++ src/plugins/walletConnectToDapps/hooks/useCallRequestEvmFees.ts
@@ -1,8 +1,10 @@
 import { fromAccountId } from '@shapeshiftoss/caip'
-import type { EvmBaseAdapter, EvmChainId, FeeDataKey } from '@shapeshiftoss/chain-adapters'
+import type { EvmBaseAdapter, EvmChainId } from '@shapeshiftoss/chain-adapters'
+import { FeeDataKey } from '@shapeshiftoss/chain-adapters'
+import { useWalletConnectState } from 'plugins/walletConnectToDapps/hooks/useWalletConnectState'
+import { assertIsTransactionParams } from 'plugins/walletConnectToDapps/typeGuards'
+import type { WalletConnectState } from 'plugins/walletConnectToDapps/types'
 import { getFeesForTx } from 'plugins/walletConnectToDapps/utils'
-import type { WalletConnectEthSendTransactionCallRequest } from 'plugins/walletConnectToDapps/v1/bridge/types'
-import { useWalletConnect } from 'plugins/walletConnectToDapps/v1/WalletConnectBridgeContext'
 import { useEffect, useMemo, useState } from 'react'
 import type { FeePrice } from 'components/Modals/Send/views/Confirm'
 import { getChainAdapterManager } from 'context/PluginProvider/chainAdapterSingleton'
@@ -10,35 +12,34 @@ import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { selectAssets, selectMarketDataById } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
-export function useCallRequestFees(
-  request: WalletConnectEthSendTransactionCallRequest['params'][number],
-) {
+export function useCallRequestEvmFees(state: WalletConnectState) {
   const [fees, setFees] = useState<FeePrice | undefined>()
+  const { chainAdapter, chainId, accountId, transaction } = useWalletConnectState(state)
 
-  const { evmChainId, wcAccountId } = useWalletConnect()
   const assets = useAppSelector(selectAssets)
 
   const feeAsset = useMemo(() => {
-    if (!evmChainId) return
-    const feeAssetId = getChainAdapterManager().get(evmChainId)?.getFeeAssetId()
+    const feeAssetId = chainAdapter?.getFeeAssetId()
     if (!feeAssetId) return null
+    // TODO(Q): this shouldn't be feeAsset, get the real asset from request
     const feeAsset = assets[feeAssetId]
     if (!feeAsset) return null
     return feeAsset
-  }, [assets, evmChainId])
+  }, [assets, chainAdapter])
   const feeAssetPrice =
     useAppSelector(state => selectMarketDataById(state, feeAsset?.assetId ?? ''))?.price ?? '0'
 
   useEffect(() => {
-    if (!wcAccountId || !request || !evmChainId) return
-    const { account: address } = fromAccountId(wcAccountId)
-    const adapter = getChainAdapterManager().get(evmChainId)
-    if (!(address && evmChainId && feeAsset && feeAssetPrice && adapter)) return
+    if (!transaction || !accountId || !chainId) return
+    const { account: address } = fromAccountId(accountId)
+    const adapter = getChainAdapterManager().get(chainId)
+    if (!(address && chainId && feeAsset && feeAssetPrice && adapter)) return
     ;(async () => {
+      assertIsTransactionParams(transaction)
       const estimatedFees = await getFeesForTx(
-        request,
+        transaction,
         adapter as unknown as EvmBaseAdapter<EvmChainId>,
-        wcAccountId,
+        accountId,
       )
 
       const initialFees: FeePrice = {
@@ -60,6 +61,7 @@ export function useCallRequestFees(
       }
       const result = (Object.keys(estimatedFees) as FeeDataKey[]).reduce<FeePrice>(
         (acc: FeePrice, key: FeeDataKey) => {
+          if (!Object.values(FeeDataKey).includes(key)) return acc
           const txFee = bnOrZero(estimatedFees[key].txFee)
             .dividedBy(bn(10).pow(feeAsset?.precision))
             .toPrecision()
@@ -74,7 +76,7 @@ export function useCallRequestFees(
       )
       setFees(result)
     })()
-  }, [wcAccountId, evmChainId, feeAsset, feeAssetPrice, request])
+  }, [accountId, chainId, feeAsset, feeAssetPrice, transaction])
 
   return { fees, feeAsset, feeAssetPrice }
 }
