+++ src/plugins/walletConnectToDapps/walletConnectReducer.ts
@@ -1,6 +1,6 @@
 import type { SessionTypes } from '@walletconnect/types'
-import type { WalletConnectAction, WalletConnectState } from 'plugins/walletConnectToDapps/v2/types'
-import { WalletConnectActionType } from 'plugins/walletConnectToDapps/v2/types'
+import type { WalletConnectAction, WalletConnectState } from 'plugins/walletConnectToDapps/types'
+import { WalletConnectActionType } from 'plugins/walletConnectToDapps/types'
 import type { PartialRecord } from 'lib/utils'
 
 export const walletConnectReducer = (
