+++ src/plugins/walletConnectToDapps/walletUtils.ts
